Reorganize code for better readability And comparer utility
