+++ backend/src/GroundTruthCuration.Core/Interfaces/Repositories/IGroundTruthRepository.cs
