+++ backend/src/GroundTruthCuration.Core/Interfaces/Repositories/IManufacturingDataDocDbRepository.cs
