+++ backend/src/GroundTruthCuration.Core/Interfaces/Repositories/IManufacturingDataRelDbRepository.cs
