+++ backend/src/GroundTruthCuration.Core/Interfaces/Services/IGroundTruthCurationService.cs
