+++ backend/src/GroundTruthCuration.Core/Interfaces/Utilities/IGroundTruthMapper.cs
